+++ react/utils/compatibility-layer.ts
@@ -19,12 +19,10 @@ type FetchMore = (options: FetchMoreOptions) => Promise<any>;
  */
 export const makeFetchMore = (
   fetchMore: FetchMore,
-  maxItemsPerPage: number,
+  page: number,
+  setPage: (page: number) => void,
 ): FetchMore => async ({ variables, updateQuery = () => {} }) => {
-  const { to } = variables;
-  const page = variables.page
-    ? variables.page
-    : Math.floor(to / maxItemsPerPage) + 1;
+  setPage(page + 1);
 
   await fetchMore({
     updateQuery: makeUpdateQuery(page),