+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Page is now a state instead of being defined by `from` and `too`
+
+### Added
+
+- GA documentation
+
 ## [0.4.2] - 2020-01-09
 
 ### Added